Operation "Coffee to Kill the Communist Party"
There are freshly ground coffee and high-grade coffee in the tent of the New China Federation. You can choose latte or cappuccino at will. You can choose all of them. Why would the Ant Gang do this? The purpose is to promote the new federation of China, but dubiously say that the Chinese Communist Party is the root of all evil. Don't be misled by other media, the so-called mainstream media, and some regimes. Only the elimination of the Communist Party is the only way out for Xiguo.